Implement type merging in O(n * (m + n)).
With ``n'' the number of types in a given CU and ``m'' the number
of types previously parsed.

Everytime a new CU is parsed, after resolving its internal types
based on their offset in the ABBREV section, merge them with the
already known types.

As soon as a duplicated type has been found, all following CTF IDs
have to be updated.  So use a pointer for referenced types, it makes
reusing of CTF IDs simpler.

Note that functions are not compared for merging.