Simple memory pool using a single free list.
A lot of data structure are required to parse the DWARF sections
and generate internal type representations.

Parsing a kernel results in 10M+ calls to malloc(3) and another 10M+
calls to free(3).  With such huge numbers, allocating and releasing
memory represents more than 50% of the execution time of the program.

This simple free-list based memory pool makes ctfconv(2) to 2x faster.

It now takes less than 2sec to convert a kernel.